+++ gpio_base.h
@@ -27,13 +27,13 @@
  extern "C" {
 #endif
 
-int init_gpio(uint8_t gpio);
-int set_gpio_dir(uint8_t gpio, uint8_t direction);
-int close_gpio(uint8_t gpio);
+int init_gpio(uint32_t gpio);
+int set_gpio_dir(uint32_t gpio, uint8_t direction);
+int close_gpio(uint32_t gpio);
 
 int set_gpio_val_by_fd(int fd, uint8_t val);
-int gpio_get_val_fd(uint8_t gpio);
-int set_gpio_val(uint8_t gpio, uint8_t val);
+int gpio_get_val_fd(uint32_t gpio);
+int set_gpio_val(uint32_t gpio, uint8_t val);
 
 #ifdef __cplusplus
  }
